Implement linting and formatting of vue files
This was a pain to setup and is not totally working.
When I hit Ctrl+S on VSCode, I still do not get the
proper formatting, but that will be enough for now.

I have also moved again the .vscode folder from root to
devolution-web folder, as it seems to cause some
trouble in IDE.